Adding standalone fuzz harness + host unit test for capsule manifest parser
capsule_manifest_parse() is the text parser behind proc.c's mandatory-
by-default stage-2 hardware isolation (capsule_manifest_load) and
capsule_store_load_manifest() -- a storage-facing parser that decides
whether a process gets isolated is exactly the kind of surface the
hard scan invariants ask for a standalone fuzz harness on.

Extracted capsule_manifest_parse() + its private helpers (c_streq,
c_starts, c_parse_u32, c_copy_n, c_name_ok, c_trim_copy, parse_range)
out of capsule_store.c into a new src/capsule_manifest_parse.c. The
original file also does WALFS I/O, so a host build/fuzz harness for it
would need every walfs_* symbol it references stubbed out; this half
is pure logic (only capsule_store.h + simd_zero) and can be compiled
standalone on the host, mirroring how dhcp_options.c was carved out of
dhcp.c for the same reason (tests/run_host_tests.py). c_strlen/c_copy/
c_err stay in capsule_store.c (still used by its other functions) and
are duplicated as static copies in the new file rather than shared, to
keep both translation units self-contained and the change low-risk.

Added:
- tests/test_capsule_manifest.c: host unit test (13 groups, 34 checks)
  covering a full valid manifest, defaults, capsule=off, malformed
  lines, bad/oversized names, unknown keys, indented-line-outside-block,
  over-capacity process count, malformed cards ranges, NULL/empty/
  degenerate inputs, and CR/LF/CRLF line endings.
- tests/fuzz_capsule_manifest.c: libFuzzer harness (LLVMFuzzerTestOneInput).
- Registered test_capsule_manifest.c in run_host_tests.py's manifest.
- tests/_corpus/ added to .gitignore (fuzzer-generated corpus, not
  committed).

Verified:
- Installed LLVM/clang (previously missing entirely -- tests/
  run_host_tests.py already required it and could not run at all
  before this).
- python tests/run_host_tests.py: all 3 suites pass (including the
  previously-unrunnable pre-existing test_picocompress/test_dhcp),
  test_capsule_manifest 34/34.
- Built and ran the actual libFuzzer harness with
  -fsanitize=fuzzer,address,undefined: 5,381,911 executions in 61
  seconds, zero crashes, zero ASAN/UBSan findings.
- Full aarch64 kernel build unaffected: capsule_store.c and
  capsule_manifest_parse.c compile standalone with zero warnings on
  the aarch64-none-elf cross-compiler; QEMU smoke 8/8, selftest
  battery 12/12 ALL PASS (unchanged from before the refactor).
