+++ .gitignore
@@ -64,4 +64,5 @@ virt_dump.dtb
 
 # host unit test build output
 tests/_build/
+tests/_corpus/
 tests/_cov/
